second commit
Added a dummy page, create_db.php that successfully created (some pretty
random) tables in MySQL.
(For this to work MySQL must be running on localhost and have a user
'webmaster' with access to a database 'webmaster' using the password
'TempPass123'. To set up that, add a user via
http://localhost/phpmyadmin.)

Tested validation of user form-input to prevent html/JS injections.